Move Repair Turbo Streams into a concern to make controller leaner.
